requirement.txt
asgiref==3.4.1
cffi==1.14.6
cryptography==3.4.8
Django==3.2.7
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
Pillow==8.3.2
pycparser==2.20
PyJWT==2.1.0
pytz==2021.1
sqlparse==0.4.1
uuid==1.30
